+++ src/evalexpr/eval/unary.c
@@ -1,24 +1,24 @@
 #define UNUSED_UNARY(Arg) Arg = Arg - 1 + 1
 
-int op_plus(int a, int b)
+long op_plus(long a, long b)
 {
     UNUSED_UNARY(b);
     return a;
 }
 
-int op_minus(int a, int b)
+long op_minus(long a, long b)
 {
     UNUSED_UNARY(b);
     return -a;
 }
 
-int op_bang(int a, int b)
+long op_bang(long a, long b)
 {
     UNUSED_UNARY(b);
     return !a;
 }
 
-int op_not(int a, int b)
+long op_not(long a, long b)
 {
     UNUSED_UNARY(b);
     return ~a;
