+++ src/evalexpr/numstack.c
@@ -8,16 +8,16 @@ s_numstack *numstack_init(void)
     s_numstack *o = malloc(sizeof (s_numstack));
     o->size_max = 8;
     o->size = 0;
-    o->nums = malloc(sizeof (int) * o->size_max);
+    o->nums = malloc(sizeof (long) * o->size_max);
     return o;
 }
 
-void numstack_push(s_numstack *s, int num)
+void numstack_push(s_numstack *s, long num)
 {
     if (s->size >= s->size_max)
     {
         s->size_max *= 2;
-        s->nums = realloc(s->nums, sizeof (int) * s->size_max);
+        s->nums = realloc(s->nums, sizeof (long) * s->size_max);
     }
     s->nums[s->size++] = num;
 }
